feat: group modals by screen
