+++ src/screens/Partnership/EditPartnershipModal/styles.ts
@@ -22,8 +22,6 @@ export const TextInput = styled.TextInput`
 `;
 
 export const AddPartnerView = styled.View`
-  justify-content: "space-between";
-  flex-direction: "row";
   gap: 140.25px;
   margin-bottom: 24px;
 `;
@@ -36,6 +34,24 @@ export const StatusView = styled.View`
   align-items: center;
   width: 345px;
   height: 56px;
+
+  justify-content: "space-between";
+  flex-direction: "row";
+  gap: 140.25;
+`;
+
+export const DropDownArea = styled.View`
+  width: 100%;
+  height: 480px;
+  border-radius: 10px;
+  margin-top: 15px;
+  background-color: #fff
+  box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
+  border: 1px solid #ccc;
+  align-self: center;
+  margin-bottom: 10px;
+  line-height: 1.5px;
+  padding: 10px;
 `;
 
 export const ClassicationDropDownArea = styled.View`
@@ -52,19 +68,16 @@ export const ClassicationDropDownArea = styled.View`
   padding: 10px;
 `;
 
-export const DropDowArea = styled.View`
-  width: 100%;
-  min-height: 100px;
-  max-height: 480px;
-  border-radius: 10px;
-  margin-top: 15px;
-  background-color: #fff;
-  elevation: 3;
+export const StatusTypeText = styled.Text`
+  width: 85%;
+  height: 40px;
+  border-bottom-width: 0.2px;
+  border-bottom-color: #8e8e8e;
   align-self: center;
-  margin-bottom: 10px;
+  padding: 3px;
 `;
 
-export const StateDropDowArea = styled.ScrollView`
+export const StateDropDowArea = styled.View`
   width: 100%;
   height: 450px;
   border-radius: 10px;
@@ -73,13 +86,4 @@ export const StateDropDowArea = styled.ScrollView`
   elevation: 3;
   align-self: center;
   margin-bottom: 10px;
-`;
-
-export const StatusTypeText = styled.Text`
-  width: 85%;
-  height: 40px;
-  border-bottom-width: 0.2px;
-  border-bottom-color: #8e8e8e;
-  align-self: center;
-  justify-content: center;
 `;
