+++ src/screens/Partnerships/AddPartnershipModal/index.tsx
@@ -1,4 +1,4 @@
-import { Button, Close, Drop, Text } from "@components";
+import { Drop, Modal, Text } from "@components";
 import {
   ClassificationSelectOptions,
   stateSelecOptions,
@@ -8,11 +8,14 @@ import { IModalPropsForm, IPartnership } from "@interfaces/partner.interface";
 import partnershipRequests from "@requests/partnership.requests";
 import { useState } from "react";
 import { Controller, SubmitHandler, useForm } from "react-hook-form";
-import { Modal, ScrollView, TouchableOpacity, View } from "react-native";
 import {
-  AddPartnerView,
+  Modal as ReactNativeModal,
+  ScrollView,
+  TouchableOpacity,
+  View,
+} from "react-native";
+import {
   ClassicationDropDownArea,
-  Container,
   DropDowArea,
   StateDropDowArea,
   StatusTypeText,
@@ -20,7 +23,7 @@ import {
   TextInput,
 } from "./styles";
 
-export function PartnershipForm({
+export function AddPartnershipModal({
   visible,
   onClose,
   closeAfterUpdate,
@@ -35,8 +38,8 @@ export function PartnershipForm({
   const [selectStates, setSelectedStates] = useState("");
   const [selectClassification, setSelectClassification] = useState("");
 
-  const [isStatusSelectOpen, setisStatusSelectOpen] = useState(false);
-  const [isStatesSelectOpen, setisStatesSelectOpen] = useState(false);
+  const [isStatusSelectOpen, setIsStatusSelectOpen] = useState(false);
+  const [isStatesSelectOpen, setIsStatesSelectOpen] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
   const [isClassificationSelectOpen, setisClassificationSelectOpen] =
     useState(false);
@@ -53,42 +56,25 @@ export function PartnershipForm({
     closeAfterUpdate();
   };
 
-  function handlerStatusPartenerSelected(statusSelectOptions: {
-    value: string;
-  }) {
-    setSelecteStatus(statusSelectOptions.value);
-    setisStatusSelectOpen(false);
+  function handleSelectStatus(status: string) {
+    setSelecteStatus(status);
+    setIsStatusSelectOpen(false);
   }
 
-  function handlerStatesPartenerSelected(statesSelectOptions: {
-    name: string;
-  }) {
-    setSelectedStates(statesSelectOptions.name);
-    setisStatesSelectOpen(false);
+  function handleSelectState(state: string) {
+    setSelectedStates(state);
+    setIsStatesSelectOpen(false);
   }
 
   return (
     <Modal
+      title="Adicionar nova parceria"
       visible={visible}
-      animationType="slide"
-      onRequestClose={onClose}
-      transparent
-    >
-      <Container>
-        <AddPartnerView
-          style={{
-            justifyContent: "space-between",
-            flexDirection: "row",
-            gap: 140.25,
-            marginBottom: 24,
-          }}
-        >
-          <Text>Adicionar nova parceria</Text>
-
-          <TouchableOpacity onPress={onClose}>
-            <Close color="#666666" />
-          </TouchableOpacity>
-        </AddPartnerView>
+      isLoading={isLoading}
+      onClose={onClose}
+      buttonTitle=" Adicionar parceria"
+      onPressButton={handleSubmit(onSubmit)}
+      content={
         <ScrollView>
           <View style={{ gap: 12 }}>
             <Text weight="500">Informações gerais</Text>
@@ -96,18 +82,16 @@ export function PartnershipForm({
             <Controller
               control={control}
               name="name"
-              rules={{
-                required: "informe o nome do parceiro",
-              }}
+              rules={{ required: "informe o nome do parceiro" }}
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Parceria</Text>
                   <TextInput
                     placeholder="The Bugger Ducks"
                     onChangeText={onChange}
                   />
                   {errors.name && <Text>Este campo é obrigatório</Text>}
-                </View>
+                </>
               )}
             />
 
@@ -115,13 +99,13 @@ export function PartnershipForm({
               control={control}
               name="email"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>E-mail</Text>
                   <TextInput
                     placeholder="nome@gmail.com"
                     onChangeText={onChange}
                   />
-                </View>
+                </>
               )}
             />
 
@@ -129,27 +113,19 @@ export function PartnershipForm({
               control={control}
               name="classification"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Classificação</Text>
                   <TouchableOpacity
-                    onPress={() => {
-                      setisClassificationSelectOpen(
-                        !isClassificationSelectOpen,
-                      );
-                    }}
+                    onPress={() =>
+                      setisClassificationSelectOpen(!isClassificationSelectOpen)
+                    }
                   >
-                    <StatusView
-                      style={{
-                        justifyContent: "space-between",
-                        flexDirection: "row",
-                        gap: 140.25,
-                      }}
-                    >
+                    <StatusView>
                       <Text>{selectClassification}</Text>
                       <Drop />
                     </StatusView>
                   </TouchableOpacity>
-                  {isClassificationSelectOpen ? (
+                  {isClassificationSelectOpen && (
                     <ClassicationDropDownArea>
                       {Object.keys(ClassificationSelectOptions).map(
                         classification => {
@@ -169,45 +145,35 @@ export function PartnershipForm({
                         },
                       )}
                     </ClassicationDropDownArea>
-                  ) : null}
-                </View>
+                  )}
+                </>
               )}
             />
 
             <Controller
               control={control}
               name="status"
-              rules={{
-                required: "informe o status da parceria",
-              }}
+              rules={{ required: "informe o status da parceria" }}
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Status</Text>
                   <TouchableOpacity
-                    onPress={() => {
-                      setisStatusSelectOpen(!isStatusSelectOpen);
-                    }}
+                    onPress={() => setIsStatusSelectOpen(!isStatusSelectOpen)}
                   >
-                    <StatusView
-                      style={{
-                        justifyContent: "space-between",
-                        flexDirection: "row",
-                        gap: 140.25,
-                      }}
-                    >
+                    <StatusView>
                       <Text>{selectStatus}</Text>
                       <Drop />
                     </StatusView>
                   </TouchableOpacity>
                   {errors.status && <Text>Este campo é obrigatório</Text>}
-                  {isStatusSelectOpen ? (
+                  {isStatusSelectOpen && (
                     <DropDowArea>
                       {statusSelectOptions.map(status => {
                         return (
                           <TouchableOpacity key={status.id}>
                             <StatusTypeText
                               onPress={() => {
-                                handlerStatusPartenerSelected(status),
+                                handleSelectStatus(status.value);
                                 onChange(status.value);
                               }}
                             >
@@ -217,8 +183,8 @@ export function PartnershipForm({
                         );
                       })}
                     </DropDowArea>
-                  ) : null}
-                </View>
+                  )}
+                </>
               )}
             />
 
@@ -229,46 +195,38 @@ export function PartnershipForm({
                 <View>
                   <Text>Estado</Text>
                   <TouchableOpacity
-                    onPress={() => {
-                      setisStatesSelectOpen(!isStatesSelectOpen);
-                    }}
+                    onPress={() => setIsStatesSelectOpen(!isStatesSelectOpen)}
                   >
-                    <StatusView
-                      style={{
-                        justifyContent: "space-between",
-                        flexDirection: "row",
-                        gap: 140.25,
-                      }}
-                    >
+                    <StatusView>
                       <Text>{selectStates}</Text>
                       <Drop />
                     </StatusView>
                   </TouchableOpacity>
 
-                  {isStatesSelectOpen ? (
-                    <Modal
+                  {isStatesSelectOpen && (
+                    <ReactNativeModal
                       onRequestClose={() =>
-                        setisStatesSelectOpen(!isStatesSelectOpen)
+                        setIsStatesSelectOpen(!isStatesSelectOpen)
                       }
                     >
                       <StateDropDowArea>
-                        {stateSelecOptions.map(states => {
+                        {stateSelecOptions.map(state => {
                           return (
-                            <TouchableOpacity key={states.UF}>
+                            <TouchableOpacity key={state.UF}>
                               <StatusTypeText
                                 onPress={() => {
-                                  handlerStatesPartenerSelected(states);
-                                  onChange(states.name);
+                                  handleSelectState(state.name);
+                                  onChange(state.name);
                                 }}
                               >
-                                {states.name}
+                                {state.name}
                               </StatusTypeText>
                             </TouchableOpacity>
                           );
                         })}
                       </StateDropDowArea>
-                    </Modal>
-                  ) : null}
+                    </ReactNativeModal>
+                  )}
                 </View>
               )}
             />
@@ -277,13 +235,13 @@ export function PartnershipForm({
               control={control}
               name="city"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Cidade</Text>
                   <TextInput
                     placeholder="São José dos campos"
                     onChangeText={onChange}
                   />
-                </View>
+                </>
               )}
             />
 
@@ -291,10 +249,10 @@ export function PartnershipForm({
               control={control}
               name="zipCode"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>CEP</Text>
                   <TextInput placeholder="12654-356" onChangeText={onChange} />
-                </View>
+                </>
               )}
             />
 
@@ -302,13 +260,13 @@ export function PartnershipForm({
               control={control}
               name="address"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Endereço</Text>
                   <TextInput
                     placeholder="Rua 21, 123"
                     onChangeText={onChange}
                   />
-                </View>
+                </>
               )}
             />
 
@@ -316,14 +274,14 @@ export function PartnershipForm({
               control={control}
               name="memberNumber"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Número de membros</Text>
                   <TextInput
                     placeholder="100"
                     keyboardType="number-pad"
                     onChangeText={onChange}
                   />
-                </View>
+                </>
               )}
             />
 
@@ -331,25 +289,19 @@ export function PartnershipForm({
               control={control}
               name="phoneNumber"
               render={({ field: { onChange } }) => (
-                <View>
+                <>
                   <Text>Telefone</Text>
                   <TextInput
                     keyboardType="phone-pad"
                     placeholder="(12)99454-3275"
                     onChangeText={onChange}
                   />
-                </View>
+                </>
               )}
             />
           </View>
         </ScrollView>
-
-        <View style={{ padding: 20 }}>
-          <Button type="filled" onPress={handleSubmit(onSubmit)}>
-            Adicionar parceria
-          </Button>
-        </View>
-      </Container>
-    </Modal>
+      }
+    />
   );
 }
